moved eat_flags from eatmemory.c to args.c where it should be
